backend/requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
boto3==1.34.35
botocore==1.34.35
certifi==2023.11.17
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.1.0
django-filter==23.2
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.16.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
Pillow==10.0.0
promise==2.3
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
Rx==1.6.3
s3transfer==0.10.0
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
urllib3==2.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
